improved on our dictionaries
got one with 5000 common words
